refine api lib and add tests
